refactor: types.tsを削除し、各コンポーネントファイル内に型定義を移動
